libft almost clear?

+++ libft/latest/42libft/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: flseaill <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/21 21:40:26 by flseaill          #+#    #+#             */
-/*   Updated: 2017/02/15 16:54:18 by flseaill         ###   ########.fr       */
+/*   Updated: 2017/04/03 12:31:29 by flseaill         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (s && f)
